PULPissimo: Represent memory addresses with uintptr_t
uintptr_t seems to be the correct type to represent memory
addresses (e.g. zephyr seems to do it like this too). It is an unsigned
integer that is guaranteed to be able to hold any pointer value. The
alternative would be a void pointer (e.g. linux kernel, more for
historical reasons afaik(?)), but uintptr_t also has the advantage from
discouraging people from dereferencing it and instead using the proper
interfaces (writew, readw).