PULPissimo: Use isr table and mcause to redirect interrupts
So far we only considered timer interrupts. By introducing a jump
table (isr_table) and index into it with mcause we can differentiate the
interrupt sources. This is a bit unfortunate because our external
interrupts are vectored, but we force them through a single interrupt
handler (freertos_risc_v_trap_handler), thus losing the information
about the source. This is a temporary solution for now. Ideally the
context switch handler is a separate assembly routine hooked up to the
timer vector entry.