PULPissimo/link: Swap private bank0 and bank1
Private bank0 now contains .text and private bank1 .data. We leave the
first 2KiB (half of the zero page) for the tiny_data section, which
allows a single load to access the data therein with the positive range
of the 12-bit immediate.

This change makes it easier for the heap to grow because now it just can
grow into the L2 shared part instead of colliding with .text. This also
make sure that .text is nearly always in the private bank which is
advantagous since the fc has no icache and thus is always fetching.