pulp: Copy Demo/RISC-V_PULPissimo for generic RISC-V_PULP_Project
Demo/RISC-V_PULPissimo is a standard FreeRTOS Demo that does not more
than run blinky and demonstrate the workflow.

This commit starts with a new directory (RISC-V_PULP_Project) in the
top-level based on Demo/RISC-V_PULPissimo. This project is meant to be
more general in the sense that we try to support most PULP based
platforms and provide more infrastructure (tests, benchmarks, drivers,
analysis scripts etc.).

Ideally we would start a new git project where we just re-use the kernel
repository, but this would complicate access to FreeRTOS-Plus things
though we don't use it yet. Furthermore transplanting history to an
unrelated repository is pretty annoying, but we probably have to do it
at some point.