treewide: Overhaul timer API and tests
* Add layer for (base) timer API
* Rename some macros to distinguish between timer0 and timer1. Timer0
uses the lower bits to count, timer1 uses the higher bits. Hence, timer0
is called TIMER0_LO, while timer1 is called TIMER1_HI
* Add test to trigger free-running counting from the base timer (LO, HI)
from the FC. Add same test when the controlling core is the cluster core
0 (master). Since the timer in this configuration is used for
benchmarking, the test is called `timer_bench`
* Rename existing timer test as `timer_tick`, where the base
timer (either LO or HI) is used as tick, an interrupt is triggered after
a configurable frequency and the timer value is reset after the
interrupt event.